+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed  
 - Fix `curly.post` and `curly.head` using wrong libcurl options to set the HTTP Method.
 - Fix `postinstall` script not working properly.
-### Added  
+### Added
+- Added back prebuilt binaries for:
+  - Electron v3, v4 and v5
 ### Changed  
 
 ## [2.1.1] - 2020-04-28