+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added back prebuilt binaries for:
   - Electron v3, v4 and v5
+- Added `isMonitoringSockets` boolean readonly property to `Easy` instances, it is `true`
+    when `monitorSocketEvents` has been called on that `Easy` instance.
 ### Changed  
 
 ## [2.1.1] - 2020-04-28